requirements.txt
python-dotenv~=1.0.0
networkx~=3.1
opencv-python~=4.7.0.72
numpy~=1.24.3
matplotlib~=3.7.1
scikit-image~=0.20.0
scipy~=1.10.1
CairoSVG~=2.7.0
sympy~=1.12
pillow~=9.5.0
requests~=2.31.0
pandas~=2.0.2
openpyxl="*"